OmniServicos/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
psycopg2-binary==2.9.7
flask-jwt-extended
werkzeug